+++ lib/player/rotation/rotation_player_extensions.dart
@@ -1,6 +1,6 @@
 import 'package:bonfire/lighting/lighting_config.dart';
 import 'package:bonfire/player/extensions.dart';
-import 'package:bonfire/player/rotation_player.dart';
+import 'package:bonfire/player/rotation/rotation_player.dart';
 import 'package:bonfire/util/collision/collision.dart';
 import 'package:flame/animation.dart' as FlameAnimation;
 import 'package:flutter/widgets.dart';