+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.6.27]
+* little performance improvement;
+* remove mandatory joystick in widget;
+* remove param gameComponent in `LightingConfig`;
+
 ## [0.6.26]
 * improvements image cache in map load by Tiled;